exporter: factor out init, simplify update
The current create or update CRD logic tracks the CR object creation
using a variable, and works fine under the assumptions that:
- the daemonset never restarts
- the daemonset owns the CRs (e.g. no external entities creates or
  deletes them).

We can make the code more robust and also simpler by refactoring the
create/update logic following the standard controller loop more closely.

Additionally, we move the once-per-process-lifetime initialization logic
outside of the loop - the hostname is unlikely to change this often, if
changes at all, and no need to (re)create a clientset each update
attempt.

Signed-off-by: Francesco Romani <fromani@redhat.com>