finder: remove now-useless guard
When working at the very beginning of the crio support
we added this guard to prevent crashes; later on we
fixed the env vars detection to work seamlessly with crio
and containerd (no need to explicitly be told which runtime we
are working with).

Due to an oversight, the guard was not removed, preventing
the device accounting to work at all under crio.

Just removing the guard let the device accounting work as
expected.

Signed-off-by: Francesco Romani <fromani@redhat.com>