rte: add smart polling
The "smart polling" idea is all about have more polling triggers than
just timer. We know -and we can pretty much depend on- that kubelet
rewrites it checkpoints for resource manager each time a resource
allocation is performed. When this happens, something changed and we
can poll again the podresources API to learn what.

This allows much faster response time without the need of having a
very frequent poll interval, which is too taxing on steady state.

To quickly learn about the checkpoint changes, we use the standard
`fsnotify` linux API using the go package.

The only drawback of this (optional) approach is that RTE now requires
RO access to the while kubelet state directory(ies).

Signed-off-by: Francesco Romani <fromani@redhat.com>