rte: podres: shared pool from designated container
Until the kubernetes issue https://github.com/kubernetes/kubernetes/issues/102190
is fixed, we can't learn easily from the podresources API which
cpus belong to the shared pool and which cpus are assigned to exclusive
pools. This makes the resource reporting inconsistent.

A possible workaround is to explicitely design a container whose
cpu resources belong to the shared pool. Knowing that, RTE can safely
compute which cpus pertain to which pool when scanning the resources
allocated to each container.

We add to manifests a simple "pause" minion container, which we know
for sure:
1. it always runs
2. where it runs (namespace, pod name, container name)
3. gets cpus from the shared pool

So we can point back RTE to it to learn about the shared pool
composition.

Signed-off-by: Francesco Romani <fromani@redhat.com>