e2e: rte: add basic smart polling e2e test
The intent is to check that an update comes quicker than the periodic update.
Since we can't predict the timing, we need to check for the annotation
in the updated object: that's the only bit of information that can tell
us for sure the reason why this update was posted.

Signed-off-by: Francesco Romani <fromani@redhat.com>