ci: deflake: tolerate missing kubeletconfig
A streak of flakes emerged lately. The e2e tests
seem to fail to collect the kubelet config from workers.
Lacking better APIs, to do that we invoke kubectl procy
using the k8s e2e framework.

This is relatively costly and complex, and we use this
data only to validate a single field in the context
of NodeResourceTopologies validation.

However, we can meaningfully validate NRTs even without this
field, and tolerating errors getting the kubelet config
reduces the test flakiness.

So, until we figure out the root cause, which can take
some time, we tolerate this error: we now just log instead
of fail the run.

Signed-off-by: Francesco Romani <fromani@redhat.com>