Improve sorting before comparing
From time to time the resourcemonitor UT failed becasue the structs are equals but in different order.
Usually this is happening because the sorting of the structs isn't good enough.
In this update we added additional sort in which even if the ResourceName is identical we'll sort by DeviceIds, thus it increase the chances for both structs to look exactly the same.
If the error will still occur we'll add sort by third field and so on.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>