resmon: simplify the handling of empty NUMA cells
In commit fcc46913 we added code to make sure to report
topology zones which have no allocatable resources (e.g. because
everything is reserved). This patch simplifies that code
while keeping the same behaviour (bar bugs).

We take the chance to add explicit test to address
the reserved CPUs scenarion, instead of conflate all
the scenarios in a single test.

Signed-off-by: Francesco Romani <fromani@redhat.com>