kubeletconfig: allow to get from /configz
Augment and reshape the existing `kubeconf` package
to allow to get the kubelet configuration from
either a local file or the API, through a proxy.

This PR consolidates the kubelet config extraction
code from this project and
https://github.com/k8stopologyawareschedwg/deployer/tree/main/pkg/kubeletconfig

A future PR in the deployer repo will just make use
of the code here.

Signed-off-by: Francesco Romani <fromani@redhat.com>