E2E test tuning for RTE d/s repo
We want to reuse the same E2E for the RTE repo under openshift-kni org.
For that we need to do small adaptations in order to be able to vendor and compile the E2E test under the openshift-kni org

In this commit we pulled out all constants from the e2e_test.go file since it's not being vendor, thus the code does not compile since it needs those constants.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>