test: e2e: split test which require local resources
The E2E testsuite has some tests which require access
to local resources (and not to the cluster), most notably
the rte binary itself.
It's good to have such tests and we should keep them.
But we should also split them from the RTE tests proper -
which run against a cluster, to make the integration
and the consumption of the e2e suite easier.

Signed-off-by: Francesco Romani <fromani@redhat.com>