manifests: update node-res-topo to 0.0.10
We need a local copy of the noderesourcetopology-api
to run the e2e tests. We forgot to update it previously,
so we do it now.

Unfortunately, the most important e2e tests can't run on the
GH CI because the machines are not enough powerful, so any
breakage is likely to have be gone undetected.

Signed-off-by: Francesco Romani <fromani@redhat.com>