e2e: cooldown after removing a gu pod
We need to run guaranteed pods requesting exclusively allocated
CPUs in our tests. Unfortunately, we cannot yet reactly quickly to pod
deletion (only to pod creation), so we can hit false negatives when
the pod lifecycle overlaps between tests.

IOW, we need to avoid that a guaranteed pod (and the resource it
consumes) spill from a test to another; but it is not just the pod
proper we need to track, but also the state the RTE sees and reports
in the NRT objects.

This means we need to have a consistent representation of the cluster
state in the NRT objects *before* a new test starts. So the following
events need to happen in this order:

1. nth test ends
2. pods associated with nth test are cleaned up; the exclusive resources
   are freed
3. the RTE polls the kubelet state (we cannot use a smart poll here)
4. the nth+1 test starts

The only way we have now to ensure this sequence of events is to wait
$POLL_INTERVAL+tiny_amount between the nth test end and the nth+1 test
start, to give time to the RTE to catch up.

Signed-off-by: Francesco Romani <fromani@redhat.com>