e2e: rte: attempt to not lose the reactive event
We have a e2e test which want to capture the update due
to a smart polling. There is however a tiny time window
on which the update can be delivered, but too fast,
so the test doesn't detect it, and then fails as false negative.

This PR introduce strictrr sync in the test code, trying
to make this event miss less likely.

Signed-off-by: Francesco Romani <fromani@redhat.com>