test: e2e: avoid hardcoded value, introspect pod
We use hardcoded value in the tests mostly (= 95% of the cases)
just because it's the simplest way. A more robust and flexible
way is to introspect the actual pod and extract the values
we care about. After all, there often is a set of correct values,
and we care more about state matching expectations (with again
a set of correct values, not a single one) rather than a exact check.

Signed-off-by: Francesco Romani <fromani@redhat.com>