rte: cleanup: proper Run()/Stop() methods
We are reviewing the internal interface of all the
packages. We want to make sure the goroutines are always
handled by the caller (not by the callee).

In order to implement the changes, we make sure all
the relevant modules expose the Run()/Stop() methods,
including the internal ResourceObserver object.

Signed-off-by: Francesco Romani <fromani@redhat.com>