notification: add EventSource type
Add notification.EventSource type, implementing the Run()/Stop()
interface, such as it complies with the ongoing refactoring,
so the goroutines are always

This cleanup and code movement also allows us to improve
how we handle the fsnotify, which should help prevent deadlocks.

In any case, the code is now tidier and easier to follow
and to test.

Resolves: https://github.com/k8stopologyawareschedwg/resource-topology-exporter/issues/93
Signed-off-by: Francesco Romani <fromani@redhat.com>