e2e:fix FindNotificationFilePath function
RTE manifest layout has changed in latest deployer version, when underthe RTE container we splitted the command and arguments of RTE container into separte fileds.
This requires the FindNotificationFilePath to be looking for the --notify-file argument in the correct place.

Furthermore, we didn't check the returned error from the function, let's fix it as well.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>