tools: add tool to generate fake load
With small additions to our existing packages,
we can build a simple NRT load generator, useful to
test, troubleshoot and assess the scalability of RTE stack.

The easiest part, which is the one addressed in this change,
is to stress test the apiserver part generating high amount
of synthetic yet realistic NRT data.

Generating fake data replacing the kubelet source is harder
and requires more infra - will be addressed in the future.

Note this tool is intentionally not shipped yet in releases:
still needs to mature a bit at least.

Signed-off-by: Francesco Romani <fromani@redhat.com>