resmonitor: update devices capacity
It is not trivial to detect devices capacity from the node as we're doing with today with CPU and memory.
Getting the real device capacity will require an access to Node.Status or device API plugin. Both way are too expensive and does not worth it, since TAS doesn't care too much about capacity, but more about allocatable and available values.
Due to the above, we decided to initialize devices` capacity same as devices` allocatable, which is the most cheaper and harmless way to handle this issue.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>