resmonitor: NewResourceMonitor friendly API
For the exsiting unit-test to pass, we should use a mock K8S client.
This means we should create `ResourceMonitor` with custom client.

Let's make the constructor friendlier and use functional options:
https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis

If case some of the fields won't be initialized, we fall back to the default values (the ones used before).

No expected changes in the behavior for this patch.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>