config: add and use the koanf package
Leveraging the 3rd party koanf package, we can achieve a longstanding goal for
RTE: really useful configuration file handling, AND 1:1 mapping between
the configuration file and the command line flags.

This mean we can have:
1. hardcoded defaults.
2. a configuration file, which can override any hardcoded default
3. the commandline flags, which can override defaults and take
   precedence over the configuration file settings.

This allow to greatly reduce the overabundance of commandline flags,
and allow shipping configuration file to simplify the deployment.

Signed-off-by: Francesco Romani <fromani@redhat.com>