e2e: gh actions: split e2e lanes
split the e2e lane in polling and evented, with different
RTE configurations: in the former, use only polling; in the latter,
use only events.
This is meant to enable more focused, thus precise, e2e tests.

A lane with both enabled is intentionally left out for now, given
the high CI cost is unclear the benefits it will give us.

Some tests which are not related to the update mechanism are
intentionally run in both scenarios. The highest cost is the
complilation of code, which is shared among lanes.

Signed-off-by: Francesco Romani <fromani@redhat.com>