notification: allow to disable the polling
Up until now, it was not possible to disable the polling.
The workaround was to set the polling interval to a insanely high
value (like 1 day) but still, this is not a real fix.

The assumption was some form of polling was always useful, but
over time this turns out to be not really true, and disabling
polling is helpful at very least in testing.

Implying some form of polling was more than a historical combination
of events than a conscious design decision, so let's fix it, better
late than never.

Signed-off-by: Francesco Romani <fromani@redhat.com>