e2e: events: send events while waiting
We should trigger the events which are expected to unleash
a reactive notification while we are already waiting for these events,
to minimize (avoid) the risk of losing that event.

We use a more correct test layout in the e2e notification hook
test, so let's move to that one.

Signed-off-by: Francesco Romani <fromani@redhat.com>