resmon: pfp: expose pfp compute method
The previously-added `--pods-fingerprint-unrestricted` was clumsy and
unexpressive. Drop it and add a more generic and expressive
`--pods-fingerprint-method` which accepts (potentially a subset of)
the same methods exposed by the podfingerprint package.

While the input interface obviously change, the produced output
is expected to be identical.

This change should not be backported to the v0.11 branch.

Signed-off-by: Francesco Romani <fromani@redhat.com>