resource-topology-exporter: filter terminal pods:
we want RTE and the scheduler to have the same PFP value.
For that they should have the same picture of exisiting
pods under the node.

But scheduler watches only non-terminal pods, while RTE compute PFP
by taking into account all the pods.

In order to overcome this issue we want to
filter the terminal pods from the podresourceAPI List call.

This is optional and disabled by default

Signed-off-by: Talor Itzhak <titzhak@redhat.com>