e2e:fixture: introduce fixture
The fixture contains context and clients that will be used
during the test run.
We're using this proven technique in other projects

The fixture initialized once in order to retain the same context.
The fixture intended to replace the k8s framework client.

We're using this technique in other projects
and it proven to be convenient, so let's
continue with this approach.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>