notification: statedirs: drop feature
Remove support to get notifications watching the kubelet state dirs.
As far as kubelet is implemented, it's hard to secure, and has issues
anyway (cannot detect deletes).
A better replacement is use a OCI hook to leverage the notification
file.

This feature has been deprecated since 0.10.z, time to remove it for
good.

Signed-off-by: Francesco Romani <fromani@redhat.com>