nrtupdater: create or get client once
instead of recreating the NRT clieant at each update (!!) from
the default kubeconfig path (!!!!), get it from the client,
optionally creating it once at startup if given nil value.

More than performance, whose gains are expected to be
negligible, we do this change to make the code more
testable and easier to consume.

Signed-off-by: Francesco Romani <fromani@redhat.com>