metrics: add support for TLS server
consume the TLS options to listen securely over TLS
when serving metrics.
Update the manifests to consume the TLS certs injected
into the pod (secrets or configmaps). The TLS-related
mounts are optional because a user may want to omit them
if they disable the metrics or expose them into http
mode for whatever reason.

Signed-off-by: Francesco Romani <fromani@redhat.com>