uplift github.com/jaypipes/ghw to v0.12.0
This commit uplifts the github.com/jaypipes/ghw dependency to the
v0.12.0 release (released July 2023). Included in the v0.11.0 release
was a fix for detecting memory areas in hardware with NUMA architecture
and Linux kernels > 5.19.0 as well as ARM platforms.

We are running resource-topology-exporter in our environments and ran
into the aforementioned bug in ghw 0.9.0, thus this commit to uplift to
a more modern ghw release.

Related https://github.com/jaypipes/ghw/issues/341

Signed-off-by: Jay Pipes <jaypipes@gmail.com>