podres: filter: add Verify functions
the Verify family of functions allow check if a given
podresources List Response matches some criteras, but
return a richer set of information useful for debug
and for better filtering decisions.

This function is a superset of the existing Required endpoint
which augments the functionality, not replaces it.

The older functionalty (Required) is however now deprecated
but still supported.

Signed-off-by: Francesco Romani <fromani@redhat.com>