resmon: per-flow verbosity for NormalizeContainerDevices
We call `NormalizeContainerDevices` both in the `Scan` flow
and in the `updateNodeAllocatable` flow. We want different
verbosity levels for them.

The `Scan` flow is done regularly, so we want to log only
with VERY high verbosiness.
The `updateNodeAllocatable` flow is done seldom, possibly once,
so we can afford and we actually welcome logs at low
verbosity.

Signed-off-by: Francesco Romani <fromani@redhat.com>
(cherry picked from commit 3cfd4bd5dffff9914154798886a5710f17659da3)