config: support `PFP_STATUS_FILE` env var
Fetching the value of an environment variable is simplier than fetching a container argument using oc exec command.Support the option of processing an environment variable in addition to the flag. The functionality of enabling podfingerprinting is still parsed as args and the exported file name will be used only when the feature is enabled.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>